Registration Endpoint Added
An endpoint for user registration has been added. This endpoint is
located at /authmngr/register/ and accepts POST requests. The request
body should be form encoded and contain the following fields:
* username
* password
* email

If the request is successful, the response will be a 201 status code
and the response body will contain the following fields:
* username
* email