Implemented Login/Logout Endpoints
Yhe users can now login and logout using the endpoints provided by the
API. The login endpoint returns a token which can be used to access
the other endpoints. The logout endpoint invalidates the token.

The login endpoint is at /authmngr/login/ and the logout endpoint
is at /authmngr/logout/. The login endpoint accepts a POST request with
the following parameters:
* username and/or email
* password

The logout endpoint accepts a GET requests and requires a token to be
passed in the header.