Implemented time series version of the centrality views
The implemented endpoints return the centrality metrics evolution
over time from the given start date to the end date. The metrics
are computed for each time frame from the start date to a sliding
date that is incremented by 120 days until the end date is reached.

The returned data includes only the doi, publication date, and the
centrality metrics for article relevant to the time frame, unlike the
response of the non time series endpoints that return the full article
data.